fix: bug that some rules are not working in .vue files
Some rules like `semi` have TypeScript-specific rules (@typescript-eslint/semi).
Such rules are not worked in *.vue files (neither `semi` nor `@typescript-eslint/semi` worked)
